Moved the api for the event service to core api
