+++ repose-aggregator/core/repose-core-api/src/main/java/org/openrepose/core/services/event/EventDispatcher.java
@@ -17,20 +17,12 @@
  * limitations under the License.
  * =_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_=_
  */
-package org.openrepose.core.services.event.common;
+package org.openrepose.core.services.event;
 
+public interface EventDispatcher {
 
-public interface EventService {
+    void dispatch();
 
-    void newEvent(Enum e, Object payload);
+    Event getEvent();
 
-    EventDispatcher nextDispatcher() throws InterruptedException;
-
-    <T extends Enum> void listen(EventListener<T, ?> el, Class<T> events);
-
-    <T extends Enum> void listen(EventListener<T, ?> el, T... events);
-
-    <T extends Enum> void squelch(EventListener<T, ?> el, Class<T> events);
-
-    <T extends Enum> void squelch(EventListener<T, ?> el, T... events);
 }
