+++ repose-aggregator/core/repose-core-api/src/main/java/org/openrepose/core/services/event/EventListener.java
@@ -17,34 +17,9 @@
  * limitations under the License.
  * =_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_=_
  */
-package org.openrepose.core.services.event.common.impl;
+package org.openrepose.core.services.event;
 
-import org.openrepose.core.services.event.common.Event;
-import org.openrepose.core.services.event.common.EventDispatcher;
+public interface EventListener<T extends Enum, P> {
 
-import java.util.Set;
-
-public class EventDispatcherImpl implements EventDispatcher {
-
-    private final Set<EventListenerDescriptor> listeners;
-    private final Event e;
-
-    public EventDispatcherImpl(Event e, Set<EventListenerDescriptor> listeners) {
-        this.listeners = listeners;
-        this.e = e;
-    }
-
-    @Override
-    public Event getEvent() {
-        return e;
-    }
-
-    @Override
-    public void dispatch() {
-        for (EventListenerDescriptor eventListenerWrapper : listeners) {
-            if (eventListenerWrapper.answersTo(e.type())) {
-                eventListenerWrapper.getListener().onEvent(e);
-            }
-        }
-    }
+    void onEvent(Event<T, P> e);
 }
