+++ repose-aggregator/core/repose-core/src/main/java/org/openrepose/core/services/config/impl/ConfigurationUpdateManagerImpl.java
@@ -26,9 +26,9 @@ import org.openrepose.commons.config.parser.common.ConfigurationParser;
 import org.openrepose.commons.config.resource.ConfigurationResource;
 import org.openrepose.commons.utils.thread.DestroyableThreadWrapper;
 import org.openrepose.commons.utils.thread.Poller;
-import org.openrepose.core.services.event.common.Event;
-import org.openrepose.core.services.event.common.EventListener;
-import org.openrepose.core.services.event.common.EventService;
+import org.openrepose.core.services.event.Event;
+import org.openrepose.core.services.event.EventListener;
+import org.openrepose.core.services.event.EventService;
 import org.openrepose.core.services.threading.ThreadingService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
