+++ repose-aggregator/core/repose-core/src/main/java/org/openrepose/core/services/deploy/ArtifactManager.java
@@ -25,9 +25,9 @@ import org.openrepose.core.container.config.ContainerConfiguration;
 import org.openrepose.core.services.classloader.ClassLoaderManagerService;
 import org.openrepose.core.services.config.ConfigurationService;
 import org.openrepose.core.services.event.PowerFilterEvent;
-import org.openrepose.core.services.event.common.Event;
-import org.openrepose.core.services.event.common.EventListener;
-import org.openrepose.core.services.event.common.EventService;
+import org.openrepose.core.services.event.Event;
+import org.openrepose.core.services.event.EventListener;
+import org.openrepose.core.services.event.EventService;
 import org.openrepose.core.services.event.listener.SingleFireEventListener;
 import org.openrepose.core.services.threading.ThreadingService;
 import org.slf4j.Logger;
