+++ repose-aggregator/core/repose-core/src/test/groovy/org/openrepose/core/services/deploy/ArtifactDirectoryWatcherTest.groovy
@@ -25,7 +25,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.TemporaryFolder
 import org.mockito.Mockito
-import org.openrepose.core.services.event.common.EventService
+import org.openrepose.core.services.event.EventService
 
 import static org.junit.Assert.assertFalse
 import static org.junit.Assert.assertTrue
